requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
addict=2.4.0=pypi_0
aliyun-python-sdk-core=2.14.0=pypi_0
aliyun-python-sdk-kms=2.16.2=pypi_0
altair=5.1.2=pypi_0
aom=3.6.0=h6a678d5_0
attrs=23.1.0=pypi_0
blas=1.0=mkl
blinker=1.6.3=pypi_0
blosc=1.21.3=h6a678d5_0
brotli-bin=1.0.9=h5eee18b_7
brotli=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpickle=2.2.1=py39h06a4308_0
colorama=0.4.6=pypi_0
contourpy=1.1.1=pypi_0
crcmod=1.7=pypi_0
cryptography=41.0.3=py39hdda0065_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pypi_0
cytoolz=0.12.0=py39h5eee18b_0
dask-core=2023.6.0=py39h06a4308_0
dav1d=1.2.1=h5eee18b_0
deprecated=1.2.14=pypi_0
einops=0.7.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.9.2=py39h06a4308_0
future=0.18.3=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.37=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth-oauthlib=1.0.0=pypi_0
google-auth=2.23.2=pypi_0
grpcio=1.59.0=pypi_0
humanize=4.8.0=pypi_0
idna=3.4=py39h06a4308_0
imagecodecs=2023.1.23=py39hc4b7b5f_0
imageio=2.31.4=py39h06a4308_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema-specifications=2023.7.1=pypi_0
jsonschema=4.19.1=pypi_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcurl=7.88.1=h251f7ec_2
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnghttp2=1.52.0=h2d74bed_1
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py39h06a4308_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=3.0.0=pypi_0
markdown=3.5=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.0=pypi_0
mdurl=0.1.2=pypi_0
mkl-service=2.4.0=py39h5eee18b_1
mkl=2023.1.0=h213fc3f_46343
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mmcv=2.0.1=pypi_0
mmengine=0.8.5=pypi_0
model-index=0.1.11=pypi_0
monai=1.2.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
nibabel=5.1.0=pypi_0
numpy-base=1.26.0=py39hb5e798b_0
numpy=1.26.0=py39h5f9d8c6_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
opendatalab=0.0.10=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openmim=0.3.9=pypi_0
openssl=3.0.11=h7f8727e_2
openxlab=0.0.26=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
packaging=23.2=pypi_0
pandas=2.1.1=pypi_0
partd=1.4.0=py39h06a4308_0
pillow=10.0.1=py39ha6cbd5a_0
pip=23.2.1=py39h06a4308_0
platformdirs=3.11.0=pypi_0
plotly=5.17.0=pypi_0
protobuf=4.24.4=pypi_0
pyarrow=13.0.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyasn1=0.5.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.19.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.16.1=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python-dateutil=2.8.2=pypi_0
python=3.9.18=h955ad1f_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytorch=1.13.1=py3.9_cuda11.7_cudnn8.5.0_0
pytz=2023.3.post1=pypi_0
pywavelets=1.4.1=py39h5eee18b_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
requests=2.28.2=pypi_0
rich=13.4.2=pypi_0
rpds-py=0.10.4=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=py39h6a678d5_1
scipy=1.11.3=py39h5f9d8c6_0
setuptools=60.2.0=pypi_0
shellingham=1.5.3=pypi_0
simpleitk=2.3.0=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
streamlit=1.27.2=pypi_0
submitit=1.5.0=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorboard=2.14.1=pypi_0
termcolor=2.3.0=pypi_0
tifffile=2023.4.12=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=py39h06a4308_0
torchaudio=0.13.1=py39_cu117
torchio=0.19.1=pypi_0
torchvision=0.14.1=py39_cu117
tornado=6.3.3=pypi_0
tqdm=4.65.2=pypi_0
typer=0.9.0=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023.3=pypi_0
tzlocal=5.1=pypi_0
urllib3=1.26.16=py39h06a4308_0
validators=0.22.0=pypi_0
watchdog=3.0.0=pypi_0
werkzeug=3.0.0=pypi_0
wheel=0.41.2=py39h06a4308_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yapf=0.40.2=pypi_0
zfp=1.0.0=h6a678d5_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
